Create WebSocketQueryStack, move relevant code into new class
